feat(cli): unify renv and kctx into single envoke command surface
Collapses internal/cli/renv and internal/cli/kctx into cmd/envoke/main.go.
All commands are now top-level subcommands of envoke with no renv/kctx prefix.

New commands: exec, yaml, load, switch, status
Shell-init now emits a single envoke() function with unified _envoke_unload_token
and _envoke_check_unload, replacing the previous dual renv/kctx token logic.

KCTX_ prefix convention for kubeconfig directives is preserved.

Closes #54
