--- internal/cli/kctx/detach_windows.go
@@ -1,5 +1,0 @@
-//go:build windows
-
-package kctx
-
-func detachFromTerminal() {}
