--- internal/cli/renv/detach_windows.go
@@ -1,5 +1,0 @@
-//go:build windows
-
-package renv
-
-func detachFromTerminal() {}
